rename monadicpool to xpool/stateful, simplify xpool
