+++ stateful/pool.go
@@ -1,4 +1,4 @@
-// The intent of monadicpool is to support stateful objects.
+// The intent of stateful is to support stateful objects.
 //
 // Different than [xpool.Pool], the monadic [Pool] handle two different generic types: S and T
 //   - T is the type of the object returned from the pool
@@ -18,7 +18,7 @@
 //
 // We can use [New] to create a monadic [Pool] that manage the state of [bytes.Reader] via Reset method implicity:
 //
-//	pool := monadicpool.New[[]byte](func() io.Reader { // you can use any interface that you want
+//	pool := stateful.New[[]byte](func() io.Reader { // you can use any interface that you want
 //	  return bytes.NewReader(nil)
 //	}
 //
@@ -26,7 +26,7 @@
 //	defer pool.Put(br)      // implicit Reset(nil)     -- the zero value of state S, []byte on this case
 //
 //	// use byte reader here as io.Reader
-package monadicpool
+package stateful
 
 import (
 	"github.com/peczenyj/xpool"
@@ -51,84 +51,33 @@ type Resetter[S any] interface {
 	Reset(state S)
 }
 
-// OnResetCallback type.
-// Will be called with a true value if the value T isa [Resetter] and was called with success.
-type OnResetCallback func(called bool)
-
-type pollConfig struct {
-	onGetResets []OnResetCallback
-	onPutResets []OnResetCallback
-}
-
-// Option type.
-type Option func(*pollConfig)
-
-// WithOnGetResetCallback is a functional option.
-// Includes one or more callbacks to be executed on object reset on Get method.
-func WithOnGetResetCallback(onGetResets ...OnResetCallback) Option {
-	return func(o *pollConfig) {
-		o.onGetResets = append(o.onGetResets, onGetResets...)
-	}
-}
-
-// WithOnPutResetCallback is a functional option.
-// Includes one or more callbacks to be executed on object reset on Put method.
-func WithOnPutResetCallback(onPutResets ...OnResetCallback) Option {
-	return func(o *pollConfig) {
-		o.onPutResets = append(o.onPutResets, onPutResets...)
-	}
-}
-
 // New is the constructor of an [Pool] for a given set of generic types S and T.
 // Receives the constructor of the type T.
-// It sets a trivial resetter, that try to convert T to [Resetter]
+// It sets a trivial resetter, T must be a [Resetter]
 // will call Reset(state S) before return the object on Get(state S)
 // will call Reset(zero value of S) before push back to the pool.
-func New[S, T any](
+func New[S any, T Resetter[S]](
 	ctor func() T,
-	opts ...Option,
 ) Pool[S, T] {
-	var c pollConfig
-
-	for _, opt := range opts {
-		opt(&c)
-	}
-
-	onGetResetter := buildDefaultResetter[S, T](c.onGetResets...)
-	onPutResetter := buildDefaultResetter[S, T](c.onPutResets...)
-
 	return newWithResetters[S, T](
 		ctor,
-		onGetResetter,
+		func(object T, state S) {
+			object.Reset(state)
+		},
 		func(object T) {
 			var zero S
 
-			onPutResetter(object, zero)
+			object.Reset(zero)
 		},
 	)
 }
 
-func buildDefaultResetter[S, T any](
-	onResets ...OnResetCallback,
-) func(object T, state S) {
-	return func(object T, state S) {
-		defaultResetter, ok := any(object).(Resetter[S])
-		if ok {
-			defaultResetter.Reset(state)
-		}
-
-		for _, onReset := range onResets {
-			onReset(ok)
-		}
-	}
-}
-
-// NewWithResetter is the constructor of an [Pool] for a given set of generic types S and T.
+// NewWithCustomResetter is the constructor of an [Pool] for a given set of generic types S and T.
 // Receives the constructor of the type T as a callback.
 // We can specify a special resetter, to be called with a zero value of S before
 // return the object from the pool.
 // Be careful, the custom resetter must be thread safe.
-func NewWithResetter[S, T any](
+func NewWithCustomResetter[S, T any](
 	ctor func() T,
 	customResetter func(object T, state S),
 ) Pool[S, T] {
@@ -148,7 +97,7 @@ func newWithResetters[S, T any](
 	onGetResetter func(object T, state S),
 	onPutResetter func(object T),
 ) Pool[S, T] {
-	pool := xpool.NewWithResetter[T](ctor, onPutResetter)
+	pool := xpool.NewWithCustomResetter[T](ctor, onPutResetter)
 
 	return &resettableMonadicPool[S, T]{
 		pool:     pool,
